Add global utility type to extract components props
This utility can be used when we need to acces prop types of another component.
I don't like to export them in the case where some other component want to reuse
them. So a utility type for extraction these types is handy.

But we need to split module definitions for `*.png` files for example, from the
`global.d.ts`, because they can't live in a module which seem to be necessary to
get the global utility type working. So, the module definitions are moved into a
separate file