Add scroll indicator
Visitors not always notice that there is more content than just the hero section.
A scroll indicator will motivate to scroll down. The indicator disappears when the
about section enter the visible area. To ensure a gentle fade out, the animation
is paused when the ireation ends, where the arrow is not visible anyway. Because
the about section is also not intersecting the viewport when scolling further down
I also need to check if the hero section is visible, this ensures a proper behavior
when scrolling back up speedy.